ranges as uint64

+++ counting/counting.h
@@ -2,11 +2,11 @@
 
 int count_numbers(uint32_t *arr, int count, struct tree_owner ctx[]);
 
-void prepare_shards(struct tree_owner ctx[], uint32_t shards_count, uint32_t shard_size);
-void destroy_shards(struct tree_owner ctx[], uint32_t shards_count, uint32_t shard_size);
+void prepare_shards(struct tree_owner ctx[], uint64_t shards_count, uint64_t shard_size);
+void destroy_shards(struct tree_owner ctx[], uint64_t shards_count, uint64_t shard_size);
 
-uint32_t seen_only_once(struct tree_owner *owner);
+uint64_t seen_only_once(struct tree_owner *owner);
 
-uint32_t aggregate_unique_numbers(struct tree_owner ctx[], uint32_t shards);
+uint64_t aggregate_unique_numbers(struct tree_owner ctx[], uint64_t shards);
 
-uint32_t aggregate_seen_only_once(struct tree_owner ctx[], uint32_t shards);
+uint64_t aggregate_seen_only_once(struct tree_owner ctx[], uint64_t shards);
